+++ src/edit.js
@@ -6,7 +6,6 @@ import fs from 'fs';
 import {
     FILEFINDER,
     SHORTCUTS,
-    FILEEXPLORER,
 } from './util/modes.js';
 import {
     calcFileFinderOutput,
@@ -14,9 +13,9 @@ import {
     createSnapshot,
     isInGit,
     getData,
-    renderScreen,
     setFileSearchOutput,
 } from './util/helper.js';
+import { render } from './util/render.js';
 import { sendKeys } from './util/sendKeys.js';
 import { handleMouseInputs } from './keybinds/mouse.js';
 
@@ -116,7 +115,7 @@ if (filePath !== undefined) {
     setFileSearchOutput(state, state.git);
     calcFileFinderOutput(state);
 }
-renderScreen(state, screen);
+render(state, screen);
 
 term.on('key', (key) => {
     state.status = '';
@@ -143,7 +142,7 @@ term.on('mouse', (name, coor) => {
 
 term.on('resize', () => {
     try {
-        renderScreen(state, screen);
+        render(state, screen);
     } catch (e) {
         term.fullscreen(false);
         console.log(e);
