+++ src/util/color.js
@@ -1,0 +1,37 @@
+const BLACK = 0;
+const GREY = 8;
+const RED = 9;
+const GREEN = 10;
+const YELLOW = 11;
+const BLUE = 12;
+const MAGENTA = 13;
+const CYAN = 14;
+const WHITE = 15;
+const DARKGREEN = 22;
+const FADEDRED = 1;
+const FADEDGREEN = 2;
+const FADEDYELLOW = 3;
+const FADEDBLUE = 4;
+const FADEDMAGENTA = 5;
+const FADEDCYAN = 6;
+const SILVER = 7;
+
+export {
+    BLACK,
+    GREY,
+    RED,
+    GREEN,
+    YELLOW,
+    BLUE,
+    MAGENTA,
+    CYAN,
+    WHITE,
+    DARKGREEN,
+    FADEDRED,
+    FADEDGREEN,
+    FADEDYELLOW,
+    FADEDBLUE,
+    FADEDMAGENTA,
+    FADEDCYAN,
+    SILVER,
+};
